Fix creation of type declarations
When we switched from an eslint specific configuration to using a "main"
tsconfig.json we omitted to specify a source directory to have proper
typescript support also in our test files.

However when transpiling the types for our npm package there is no need
to bundle any test files with it.

We now explicitily set the include parameter in our
`tsconfig.types.json`