Prepare 4.0.0 Release
We bumped the `@vitest/expect` dependency to a new major version.

While the major version bump of `@vitest/expect` seems to
come from `vite` itself and no significant changes of the interface
of `@vitest/expect` seem to have happened, we sill want to follow
semver.

And forcing a new major version to our library consumers is technically
a breaking change in my opinion.