maint: Install aws-sdk-client-mock as peerDependency
Our code does depend on aws-sdk-client-mock but we have to integrate
with whatever version the consumer of our library installed. As a result
we should really list `aws-sdk-client-mock` as a peer dependency and not
as a regular dependency.

The thing we are most dependent from are `clientName` (available since
`v1.0.0`), `commandCalls`, and `calls` (available since `v0.5.6`).

For proper typing we depend on AwsCommand (since `v1.0.0`) and AwsStub
(which changed its generic type interface in `v2.2.0`

As a result this looks anything from 2.2.0 should work.