Change @smithy/types to a peer dependency
Our library does not really depend on any specific version of
smithy/types, we have to use whatever the environment dictates (e.g.
mostly depending on which version of the AWS SDK is used inside the
consumers project).

As result we should not install @smithy/types on our own as it does not
make sense to ever install a separate version from the main version used
inside the project.

While we use types from the package in our function signatures
(`MetadataBearer`), our code actually does do not depend on a specific
shape of the type. As a result we are confident in our open-ended
version constraint.