maint: Add tests for asymmetric matchers
Whenever compare call arguments (e.g. `toHaveReceivedCommandWith` we
internally use ObjectContaining and `asymmetricMatch` of `vitest`. This
is not only to allow partial matches, but also in order to use something
like `expect.any(String)` or `expect.stringMatching()`.

This now adds some tests to verify this does actually work as intended.