Do not rely on correct this
In the past we used `MatcherState` inside formatCalls which is the
`this` value of our expectation functions.

In order to pass `this` correctly we then used arrow functions for the
returned message function, so this is not overwritten.

We now solve this differently by deconstructing this (which we did
already) to also destruct `utils` and then simply pass the deconstructed
`utils` to `formatCalls`.

This removes some surprise and hopefully makes the code easier to reason
about.