Add vitest peerDependency
Our code in `extend.ts` depends on `vitest` since it extends the
`vitest` module.

The code should work with any `vitest` greater `3.2.0` according to the
[documentation][1].

Adding it as a `dependency` does not make sense since we want to augment
the installation of the user, not bring in the package on our own.

[1]: https://vitest.dev/guide/extending-matchers.html